requirements.txt
streamlit==1.32.0
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
fpdf==1.7.2


